Fixed several power and water bugs. Also implemented the CreateTechJob. Besides cleanup of old miscellaneous TODOs, I think it's time for a new task... Placing walls, perhaps?

+++ trunk/RTS/Assets/Scripts/Units/UnitStatus.cs
@@ -9,7 +9,7 @@ public class UnitStatus : ControllableStatus {
 	// Unit loses this much water on each water tick
 	public int waterLossRate;
 	public int Water { get; protected set; }
-	public float WaterPercentage { get { return Water / maxWater; } }
+	public float WaterPercentage { get { return Water / (float)maxWater; } }
 	
 	// Amount of time since last water tick
 	protected float waterTickTimer = 0;
